# openapi.yaml
openapi: 3.1.0
info:
  title: Insight Store — Telegram Forwarding Bot
  description: Automatic Telegram forwarding between authorized chats. No quotas,
    no subscription limits — and no evasion of Telegram's own limits.
  version: 0.1.0
paths:
...
  /api/v1/telegram/connections/{connection_id}:
    get:
      tags:
      - connections
      summary: Get Connection
      operationId: get_connection_api_v1_telegram_connections__connection_id__get
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/telegram/connections/bot:
    post:
      tags:
      - connections
      summary: Create Bot Connection
      operationId: create_bot_connection_api_v1_telegram_connections_bot_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotConnectionRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/telegram/connections/user/start:
    post:
      tags:
      - connections
      summary: Start User Connection
      operationId: start_user_connection_api_v1_telegram_connections_user_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartUserConnectionRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/telegram/connections/user/verify:
    post:
      tags:
      - connections
      summary: Verify User Code
      operationId: verify_user_code_api_v1_telegram_connections_user_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/telegram/connections/{connection_id}/2fa:
    post:
      tags:
      - connections
      summary: Submit 2Fa
      operationId: submit_2fa_api_v1_telegram_connections__connection_id__2fa_post
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/telegram/connections/{connection_id}/sync:
    post:
      tags:
      - connections
      summary: Sync Chats
      description: 202 and returns. The API never waits for a Telegram round trip.
      operationId: sync_chats_api_v1_telegram_connections__connection_id__sync_post
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/telegram/connections/{connection_id}/disconnect:
    post:
      tags:
      - connections
      summary: Disconnect
      operationId: disconnect_api_v1_telegram_connections__connection_id__disconnect_post
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Connection Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisconnectRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AcceptedResponse:
      properties:
        status:
          type: string
          const: accepted
          title: Status
          default: accepted
        task_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Task Id
        message:
          type: string
          title: Message
          default: Queued. Refresh to see the result.
      type: object
      title: AcceptedResponse
      description: 202 body for anything that enqueues background work.
...
    CapabilitiesResponse:
      properties:
        can_read_subscribed_channels:
          type: boolean
          title: Can Read Subscribed Channels
        can_read_group_messages:
          type: boolean
          title: Can Read Group Messages
        can_read_history:
          type: boolean
          title: Can Read History
        max_download_bytes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Download Bytes
        notes:
          items:
            type: string
          type: array
          title: Notes
      type: object
      required:
      - can_read_subscribed_channels
      - can_read_group_messages
      - can_read_history
      - max_download_bytes
      title: CapabilitiesResponse
...
    ConnectionResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        kind:
          type: string
          enum:
          - bot
          - user
          title: Kind
        label:
          type: string
          title: Label
        status:
          type: string
          title: Status
        telegram_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Telegram Username
        telegram_account_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Telegram Account Id
        last_health_check_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Health Check At
        last_successful_check_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Successful Check At
        last_error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error Code
        last_error_message_safe:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error Message Safe
        created_at:
          type: string
          format: date-time
          title: Created At
        capabilities:
          anyOf:
          - $ref: '#/components/schemas/CapabilitiesResponse'
          - type: 'null'
      type: object
      required:
      - id
      - kind
      - label
      - status
      - created_at
      title: ConnectionResponse
    CreateBotConnectionRequest:
      properties:
        label:
          type: string
          maxLength: 120
          minLength: 1
          title: Label
        bot_token:
          type: string
          maxLength: 256
          minLength: 20
          title: Bot Token
      type: object
      required:
      - label
      - bot_token
      title: CreateBotConnectionRequest
    DisconnectRequest:
      properties:
        revoke:
          type: boolean
          title: Revoke
          default: false
      type: object
      title: DisconnectRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StartUserConnectionRequest:
      properties:
        label:
          type: string
          maxLength: 120
          minLength: 1
          title: Label
        phone:
          type: string
          pattern: ^\+[1-9]\d{6,14}$
          title: Phone
      type: object
      required:
      - label
      - phone
      title: StartUserConnectionRequest
...
    TwoFactorRequest:
      properties:
        password:
          type: string
          maxLength: 256
          minLength: 1
          title: Password
      type: object
      required:
      - password
      title: TwoFactorRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerifyCodeRequest:
      properties:
        connection_id:
          type: string
          format: uuid
          title: Connection Id
        code:
          type: string
          maxLength: 16
          minLength: 3
          title: Code
      type: object
      required:
      - connection_id
      - code
      title: VerifyCodeRequest